filtered_requirements.txt
ai21==1.0.7
arxiv==1.4.7
authlib
bidict==0.22.0
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
commonmark==0.9.1
cryptography==38.0.4
cssselect==1.1.0
dataclasses-json==0.5.9
dateparser==1.1.8
dill==0.3.5.1
diskcache==5.6.1
faiss-cpu==1.10.0
Flask==3.0.2
Flask-Caching==2.0.2
Flask-Dance==7.1.0
Flask-Session==0.5.0
Flask-Limiter==3.5.0
flask-cors==5.0.1
flask-sock>=0.7.0
google-auth-oauthlib==0.4.6
goose3[all]==3.1.16
huggingface-hub==0.14.1
trafilatura==1.6.1
readabilipy @ git+https://github.com/alan-turing-institute/ReadabiliPy.git@8905cb9c22d2f813dabe86be5e9e4ea4e1f36df3
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainplus-sdk==0.0.20
langgraph==0.0.54
langchainhub==0.1.15
langchain_experimental==0.0.59
llama-index==0.10.37
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.37.post1
llama-index-llms-openai==0.1.19
llama-index-multi-modal-llms-openai==0.1.6
lxml==4.9.2
mmh3==4.0.0
more-itertools==8.14.0
oauthlib==3.2.2
openai==1.69.0
openpyxl==3.1.2
pdfminer.six==20221105
pandas==1.4.2
Pillow==9.5.0
pygpt4all==1.1.0
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pypdf==3.9.1
ptyprocess>=0.7.0
requests-oauthlib==1.3.1
semanticscholar==0.8.1
sentence-transformers==2.2.2
tenacity==8.2.2
tiktoken==0.6.0
unstructured==0.6.10
unstructured-inference==0.4.4
wikipedia==1.4.0
requests-oauthlib==1.3.1
playwright==1.32.0
pytest-playwright==0.3.2
rank_bm25==0.2.2
google-api-python-client==2.93.0
PyMuPDF==1.22.3
selenium==4.10.0
spacy==3.6.0
spacy-lookups-data==1.0.3
text-generation==0.6.0
tokenizers>=0.13.3
transformers>=4.31.0
ujson==5.8.0
vllm>=0.2.0
lru-dict==1.2.0
numpy==1.25.1
pydub==0.25.1
youtube-transcript-api==1.0.3
yt-dlp==2025.3.27
imagehash==4.3.2
ffmpeg-python==0.2.0
ipykernel==6.29.4
ipython==9.0.2
scikit-image==0.25.2
PyMuPDF==1.25.3
markdown-it-py==3.0.0
mdit-py-plugins==0.3.0
markdown-pdf==1.7
reportlab==4.4.5
img2pdf==0.6.3
mcp[cli]>=1.12.0
PyJWT>=2.8.0
